abort test dtype=float16

+++ test/test_vllm_abort.py
@@ -34,7 +34,8 @@ async def test_vllm_abort():
             gpu_memory_utilization=0.3,  # 使用较少内存
             tensor_parallel_size=1,
             trust_remote_code=True,
-            enable_chunked_prefill=False
+            enable_chunked_prefill=False,
+            dtype='float16',
         )
         
         # 2. 创建引擎
